requirements.txt
PySide6>=6.6.0,<6.8
mss>=9.0.1,<10
pytesseract>=0.3.10,<0.4
pyautogui>=0.9.54,<0.10
pynput>=1.7.6,<2
Pillow>=10.0.0,<11
pytest>=7.4.0,<9
